+++ spec/rubocop/cop/style/safe_navigation_spec.rb
@@ -35,6 +35,10 @@ describe RuboCop::Cop::Style::SafeNavigation, :config do
       expect_no_offenses('foo && foo[:bar]')
     end
 
+    it 'allows an object check before a negated predicate' do
+      expect_no_offenses('foo && !foo.bar?')
+    end
+
     it 'allows method calls that do not get called using . safe guarded by ' \
       'an object check' do
       expect_no_offenses('foo + bar if foo')